+++ packages/playground/src/app/app.component.html
@@ -1,188 +1,1 @@
-<div class="playground">
-  <div class="playground__header" [ngClass]="{'playground__header--slim': hasResults || isLoading, 'playground__header--large': !hasResults && !isLoading}">
-
-<!--    todo -->
-<!--    <label style="float: right;">-->
-<!--      <input type="checkbox"-->
-<!--             [checked]="true"-->
-<!--             (change)="help = !help"-->
-<!--             class="tooltip bottom"-->
-<!--             aria-label="Show additional explanation">-->
-<!--      Help-->
-<!--    </label>-->
-
-    <h1>Transform any Website to a RSS/ATOM/JSON Feed</h1>
-
-    <div class="form">
-      <ng-form #form (ngSubmit)="parseFromUrl()">
-        <div style="display: inline-block; text-align: left">
-          <div class="tooltip bottom" aria-label="The URL you want to RSSify">
-            <input [(ngModel)]="url"
-                   type="url"
-                   autofocus
-                   name="url"
-                   [autofocus]="true"
-                   autocomplete="off"
-                   required="true"
-                   (keyup.enter)="parseFromUrl()"
-                   placeholder="Enter the Url of a Website">
-          </div>
-          <div class="tooltip bottom" aria-label="The desired output format of the feed">
-            <select [(ngModel)]="options.output"
-                    name="output">
-              <option [ngValue]="output"
-                      *ngFor="let output of outputs">{{ output }}</option>
-            </select>
-          </div>
-    <!--      <label class="tooltip bottom" aria-label="Toggle the feed panel" *ngIf="showDebugger">-->
-    <!--        <input type="checkbox" [checked]="showFeed" (change)="showFeed = !showFeed">-->
-    <!--        Feed-->
-    <!--      </label>-->
-    <!--      <label class="tooltip bottom" aria-label="Toggle the article details panel" *ngIf="showDebugger">-->
-    <!--        <input type="checkbox" [checked]="showArticles" (change)="showArticles = !showArticles">-->
-    <!--        Articles-->
-    <!--      </label>-->
-
-          <button (click)="parseFromUrl()"
-                  [disabled]="isLoading">
-            Explore Feed
-            <span class="icon-rss"></span>
-          </button>
-
-<!--          <div>Advanced Settings</div>-->
-          <div *ngIf="!isLoading && !hasResults && history.length > 0">
-            <p><strong>Previous Queries</strong></p>
-            <p *ngFor="let url of history">
-              <a href="javascript:void(0)" (click)="parseFromHistoryUrl(url)">{{ url }}</a>
-            </p>
-          </div>
-        </div>
-      </ng-form>
-    </div>
-
-    <div class="feed-url tooltip bottom"
-         aria-label="The Feed URL that can be used in a feed reader of your choice"
-         *ngIf="isGenerated">
-      <span class="tooltip" aria-label="The Feed Url, which you can use in a feedreader of choice">Feed Url: </span>
-      <a [href]="getFeedUrl()"> {{ getFeedUrl() }}</a>
-<!--      <a href="javascript:void(0)">Harvest</a>-->
-    </div>
-  </div>
-
-  <div style="display:flex;justify-content:center; align-items:center;"
-       *ngIf="isLoading">
-    <div class="spinner"></div>
-  </div>
-
-  <div *ngIf="error">
-    {{error}}
-  </div>
-
-  <div class="playground__buttons" *ngIf="hasResults">
-
-    Input:
-    <button [ngClass]="{'inactive':!showMarkup}" (click)="showMarkup = !showMarkup">Markup</button>
-
-    Output:
-    <button [ngClass]="{'inactive':!showConsole}" (click)="showConsole = !showConsole">Console</button>
-    <button [ngClass]="{'inactive':!showFeed}" (click)="showFeed = !showFeed">Feed</button>
-    <button [ngClass]="{'inactive':!showArticles}" (click)="showArticles = !showArticles">Article Candidates</button>
-
-  </div>
-
-  <div class="playground__panels"
-       *ngIf="!isLoading">
-
-    <!-- Markup ----------------------------------------------------------------------------------------------- -->
-
-    <app-panel id="editor"
-               name="Markup"
-               (hide)="showMarkup = false"
-               tooltip="The static HTML-data used by the feed parser. Some websites are generated using JavaScript, which is not supported by rss-proxy"
-               *ngIf="showMarkup">
-      <button (click)="parseHtml()">
-        Refresh from HTML
-      </button>
-      <textarea [(ngModel)]="html"
-                placeholder="HTML markup of website">
-
-      </textarea>
-    </app-panel>
-
-
-    <!-- Console ----------------------------------------------------------------------------------------------------- -->
-
-    <app-panel id="console"
-               name="Console"
-               tooltip="Console logs from the parser"
-               (hide)="showConsole = false"
-               *ngIf="showConsole">
-      <pre style="overflow: auto; height: 77vh">
-        <span *ngFor="let log of logs">{{ log }}</span>
-      </pre>
-    </app-panel>
-
-    <!-- Feed -------------------------------------------------------------------------------------------------------- -->
-
-    <app-panel id="feed"
-               style="overflow: auto"
-               name="Feed"
-               tooltip="the articles translated in a feed in the format of choice"
-               (hide)="showFeed = false"
-               *ngIf="showFeed">
-      <select [(ngModel)]="options.output">
-        <option [ngValue]="output"
-                *ngFor="let output of outputs">
-          {{ output }} Feed
-        </option>
-      </select>
-
-<!--      <select [(ngModel)]="options.content">-->
-<!--        <option [ngValue]="pageResolution"-->
-<!--                *ngFor="let pageResolution of pageResolutions">-->
-<!--          {{ pageResolution }}-->
-<!--        </option>-->
-<!--      </select> - {{ optionsFromParser.content }}-->
-
-      <div *ngIf="feeds?.length > 0">
-        <p>Feeds Found in DOM:</p>
-        <ul>
-          <li *ngFor="let feed of feeds"><a [href]="feed.url">{{ feed.name }}</a></li>
-        </ul>
-      </div>
-
-      <pre [innerText]="feedData"></pre>
-    </app-panel>
-
-    <!-- Articles -------------------------------------------------------------------------------------------------- -->
-
-    <app-panel id="articles"
-               style="overflow: auto;"
-               name="Article Candidates"
-               tooltip="Choose a group of article that should be included in the feed"
-               (hide)="showArticles = false"
-               *ngIf="showArticles">
-      <select (change)="applyRuleFromEvent($event)" [(ngModel)]="currentRule">
-        <option *ngFor="let rule of rules"
-                [ngValue]="rule">
-          Candidate {{ formatScore(rule?.score) }} pts / {{ rule?.articles?.length }} articles / {{ rule?.id }}
-        </option>
-      </select>
-      <div>
-        <div>Score: {{ formatScore(currentRule?.score) }} pts, {{ articles?.length }} articles</div>
-        <div class="rule__selectors">Selectors:
-          <div>Context: {{currentRule?.rule.contextElementPath }}</div>
-          <div>Title: {{currentRule?.titlePath }}</div>
-          <div>Link: {{currentRule?.rule.linkPath }}</div>
-        </div>
-        <div></div>
-        <pre [innerText]="getArticles()" style="height: 60vh"></pre>
-      </div>
-    </app-panel>
-
-  </div>
-
-  <div class="playground__footer">
-    <a href="https://gitlab.com/damoeb/rss-proxy">rss-proxy</a> v. {{ getVersions().version }}@{{ getVersions().revision }} build on {{ getBuildDate() }}
-  </div>
-</div>
+<app-playground></app-playground>
