--- packages/playground/src/app/app.component.scss
@@ -1,112 +1,0 @@
-.playground {
-  height: 100vh;
-  display: flex;
-  background-color: #feffee;
-  flex-direction: column;
-
-  &__footer {
-    text-align: center;
-    padding-top: 5rem;
-    padding-bottom: 1rem;
-  }
-
-  &__header {
-    &--large {
-      h1 {
-        margin-top: 10rem;
-        text-align: center;
-      }
-      .form {
-        text-align: center;
-      }
-    }
-    &--slim {
-      h1 {
-        display: none;
-      }
-    }
-  }
-}
-button.inactive {
-  background: transparent;
-  border: 1px solid black;
-}
-a {
-  color: black;
-  border-bottom: 2px solid black;
-}
-a:hover {
-  text-decoration: none;
-  border-bottom-style: hidden;
-}
-select:focus,
-select:hover,
-input[type=text]:focus,
-input[type=text]:hover,
-input[type=checkbox]:focus,
-input[type=checkbox]:hover {
-  background-color: white;
-}
-.playground__header {
-  padding: 10px;
-}
-.playground__buttons {
-  padding: 10px;
-  text-align: center;
-}
-.playground__panels {
-  display: flex;
-  max-height: 80vh;
-  flex: 1;
-}
-app-panel {
-  flex: 1;
-}
-#console {
-  white-space: pre-line;
-}
-textarea {
-  width: 98%;
-  height: 70vh;
-}
-.feed-url {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 80vw;
-  padding-top: 15px;
-  font-size: 1.2em;
-  padding-bottom: 15px;
-}
-pre {
-  margin-left: 0;
-  margin-right: 0;
-  height: 70vh;
-  white-space: break-spaces;
-  word-break: break-all;
-}
-
-pre span {
-  display: block;
-  line-height: 1.5rem;
-  counter-increment: line;
-}
-pre span:before {
-
-  content: counter(line);
-  display: inline-block;
-  border-right: 1px solid #ddd;
-  padding: 0 .5em;
-  margin-right: .5em;
-  color: #888
-}
-.rule__selectors div {
-  padding-left: 1em;
-}
-app-panel {
-  padding: 5px;
-  border-top: 1px solid #b5b792;
-  border-left: 1px solid rgb(204, 204, 204);
-  display: flex;
-  flex-direction: column;
-}
