+++ packages/playground/src/app/components/playground/playground.component.spec.ts
@@ -1,0 +1,19 @@
+import { TestBed, async } from '@angular/core/testing';
+import { PlaygroundComponent } from './app.component';
+import {AppModule} from './app.module';
+
+describe('AppComponent', () => {
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [AppModule]
+    }).compileComponents();
+  }));
+
+  it('should create the app', () => {
+    const fixture = TestBed.createComponent(PlaygroundComponent);
+    const app = fixture.debugElement.componentInstance;
+
+    expect(app).toBeTruthy();
+  });
+
+});
