+++ packages/playground/src/environments/build.ts
@@ -1,6 +1,6 @@
 
 export const build = {
   version: '0.1.0',
-  revision: '16c6987',
-  date: '1590146164932'
+  revision: 'be7606b',
+  date: '1590954087235'
 };
