+++ packages/proxy/src/app.ts
@@ -1,5 +1,6 @@
 import * as express from 'express';
 import * as bodyParser from 'body-parser';
+import * as cors from 'cors';
 
 import logger from './logger'
 import { feedEndpoint } from './endpoints/feedEndpoint';
@@ -7,13 +8,14 @@ import {config} from './config';
 import {readerEndpoint} from './endpoints/readerEndpoint';
 
 // see http://patorjk.com/software/taag/#p=display&f=Chunky&t=rss%20proxy
-console.log(`                                                    
+console.log(`
 .----.-----.-----.    .-----.----.-----.--.--.--.--.
 |   _|__ --|__ --|    |  _  |   _|  _  |_   _|  |  |
 |__| |_____|_____|    |   __|__| |_____|__.__|___  |
                       |__|                   |_____|\n\n`);
 logger.info(`Starting rss-proxy v. ${config.build.version}@${config.build.revision}`);
 logger.info(`https://gitlab.com/damoeb/rss-proxy\n`);
+logger.info(`env: ${config.env}`);
 
 // -- express
 const app = express();
@@ -31,6 +33,10 @@ export interface ErrorsResponse {
 
 // -- endpoints
 
+if (config.env !== 'prod') {
+  logger.info('Enabling cors');
+  app.use(cors());
+}
 app.use('/', express.static('static'));
 
 readerEndpoint.register(app);
@@ -45,4 +51,6 @@ app._router.stack.forEach((route:any) => {
 });
 
 // startup
-app.listen(config.port, () => logger.info(`Listening on port ${config.port}`));
+app.listen(config.port, () => {
+  logger.info(`Listening on port ${config.port}`);
+});
