+++ packages/proxy/src/endpoints/settingsEndpoint.ts
@@ -1,0 +1,11 @@
+import {Express, Request, Response} from 'express';
+
+export const settingsEndpoint = new class SettingsEndpoint {
+  register(app: Express) {
+
+    app.get('/api/settings', (request: Request, response: Response) => {
+      response.json({message: ''});
+    });
+  }
+
+};
